Wait for socket to be created
If the socket does not exist at startup, wait for it to be created, and
connect to it as soon as it is created.

This works in that using it with Firefox and the browser extension works
fine. But it does not improve the experience: the extension has a
timeout, and if there's no response during before that, it assumes
failure and un-registers the response listener. Therefore, when the
socket later appears, its response isn't handled properly by the
extension.

For this to **really** make a difference, the browser extension should
not assume that the proxy has exited on timeout, and it should keep the
response listener connected. By doing so, when the socket is created at
any later time, the proxy will connect automatically, and the browser
extension would get a valid response, switching to a "ready" state
without the requiring any user intervention.